feat: migra valores monetários para BigDecimal, adiciona endpoints de administração de pedidos e prepara deploy
Corrige bugs de checkout, CORS e integridade encontrados em testes manuais;
adiciona endpoint de listagem de pedidos para o admin; parametriza URL do
banco, URL de imagens e porta via variáveis de ambiente; adiciona Dockerfile
para deploy em serviços baseados em container.

+++ src/main/java/com/joaoneto/ecommerce/dtos/ItemDoCarrinhoDTO.java
@@ -5,6 +5,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
@@ -20,8 +22,8 @@ public class ItemDoCarrinhoDTO {
     private Integer quantidade;
 
     @Schema(description = "Percentual de desconto aplicado ao item", example = "10.0")
-    private double desconto;
+    private BigDecimal desconto;
 
     @Schema(description = "Preço unitário do item com o desconto aplicado, válido enquanto o item está no carrinho (é um valor provisório — pode mudar até a compra ser finalizada)", example = "3149.91")
-    private double precoComDesconto;
+    private BigDecimal precoComDesconto;
 }
